fix(keyboard): Add Mac devtools binding
MacOS uses a separate devtools binding: Ctrl+Alt+I, compared to the rest
of the world who use Ctrl+Shift+I. That means we need a second binding
to properly handle opening devtools on MacOS.

Additionally, our Mac devtools binding should not be available on any
other platform and there's also no need for the regular devtools binding
to be available on Mac. Therefore, we can extend the keyboard shortcuts
system to let us specify some platforms.

The apps are easy to detect and may want different bindings (e.g. iOS
will want some MacOS bindings such as the emoji picker, but won't want
the devtools binding), so I've included them as separate platforms.

Unlike the modifier key bindings which require the exact modifiers to be
pressed, it doesn't really make sense to make a binding that only works
when you're on Linux && Windows, so I've made the platform specificr act
as an "or" instead.

Signed-off-by: Skyler Grey <skyler.grey@collabora.com>
Change-Id: I4f1f691f5e12a2e219cdec494e868099889823db