wsd: force saving before unloading if the last save had failed
There is a corner-case, where we can timeout on
saving and mark the last save as failed, while
there is nothing to save. When this happens
during unloading, we can end up wanting to
save (because the last save had failed), but
there is nothing to save--and we are stuck.

By forcing to save if the last save had failed,
we can properly update the save state and
move on.

Change-Id: Ibad378205e4bd54044a2b7f072e2f99cfd8849d0
Signed-off-by: Ashod Nakashian <ashod.nakashian@collabora.co.uk>