fix(multimedia): Insert multimedia at correct size
Previously we were not getting the size of multimedia we were inserting
as we can't use gstreamer to do this job in core and we had no process
for doing this in online. That meant that all videos were inserted with
a fallback size, which happened to be small and square. As not all
videos are small and square, this led to some videos having a box with a
considerably different aspect ratio than the video itself.

This patch is probably not quite right - for instance if we insert a
very large video it will end up much, much bigger than the slide - but
it's probably a good jumping-off point to do nicer things.

Adapt patch to LibreOffice Online.

Signed-off-by: Skyler Grey <skyler.grey@collabora.com>
Change-Id: Ica6869ddce9e7d9c89d7ff5755d26d04733b6eec