make run mode fd leak
To reproduce: open a document through a debug link and close it to
establish a baseline.

On each subsequent open and close:
$ ls /proc/`pidof coolwsd`/fd|wc
increments by 2

in ClientRequestDispatcher::handleIncomingMessage closeConnection is the
default 'false' for this request and nothing otherwise explictly closes
the socket on that side.

Anything similar is not seen in release mode with, e.g. a nextcloud instance.

Signed-off-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Change-Id: I927138ba77f54ff6ae86a7c10f003abd89549653