Allow the current area to extend beyond the visible area
By expanding the current area, we have a ready-to-render buffer around
the visible area. This improves the scrolling experience by allowing the
area immediately around the screen to be decompressed and disallowing it
to be GC'd so you don't see tiles drawing in as often.

Adapt patch to LibreOffice Online.

Signed-off-by: Chris Lord <chris.lord@collabora.com>
Change-Id: Ib0fa607acd22130aad5d4a3600d6891e1d9d8ce1