Impress: remove legacy CSS, map's shadow
Before this commit a shadow was visible making it look like the slide
sorter has a thick border. When closed, the shadow is still
there, this time in the left edge of the screen.

In the past we were using box-shadow applied to map to mimic a border
on slide-sorter, this is not needed anymore.

Remove it.

Adapt patch to LibreOffice Online.

Signed-off-by: Pedro Pinto Silva <pedro.silva@collabora.com>
Change-Id: I254a51992aff3434a563dfbfdde40bca42d3d917