esign: clean up two unused members in ESignature
availableProviderIDs doesn't have to be a member since commit
4d42b1e3e47ad0183649678d2aa931b44c5ffe10 (cool#10630 doc electronic
sign: return dialog result without an index, 2024-12-17).

availableCountryCodes was write-only since the same commit.

And fix the remaining lints in the file, so strict mode can be enabled
here.

Actual esign with the config from
<https://github.com/CollaboraOnline/online/issues/10630#issuecomment-2606548860>
continues to work.

Adapt patch to LibreOffice Online.

Signed-off-by: Miklos Vajna <vmiklos@collabora.com>
Change-Id: I55a023261bc51f69150426c6fdb4fa8a5dc4f777