Don't share ownership of Socket with RequestVettingStation
Reorganize so that the SocketPoll remains in the sole ownership of its
current SocketPoll until it can be transferred to ownership of the
DocumentBroker SocketPoll.

With parallel CheckFileInfo it might not be possible to create the
DocumentBroker, and transfer ownership of the Socket to its poll, at the
ideal ClientRequestDispatcher::handleClientWsUpgrade opportunity.

If so, wait until the CheckFileInfo results are available before
creating the DocumentBroker and doing the Socket transfer to its
SocketPoll.

Use new SocketPoll method to queue a transfer of a Socket from it to
another poll which createClientSession can use to trigger the transfer
when both the CheckFileInfo and socket are available.

Signed-off-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Change-Id: I4cbb141904faefc0915ff4de560c8e475461207d