cool#11617 browser, notebookbar: ignore '' -> Text context transition for text documents
In Writer, if you log the context changes when the cursor moves from
normal document body text to an in-table position, you'll see a '' ->
Text transition for the context, then a Text -> Table one.

The final goal for the issue will be to not do any switches, but
technically we have a <some tab> -> Home -> Table chain, so first just
fix the '' -> Text context change, given that the default tab seems to
be the Home tab (see `HOME_TAB_ID` at the top of
Control.Notebookbar.js), so no need to switch there.

Signed-off-by: Miklos Vajna <vmiklos@collabora.com>
Change-Id: If4a9e76320a20f09e7ee8913187a703a3c4b8249