cool#11617 browser, notebookbar: don't automatically switch from Review for text documents
Now that changing from text context to table context is a single
notebookbar switch, further decouple actions:

- showing the previously hidden Table tab is fine

- switching to that tab may be problematic

Keep the current behavior by default (so if you e.g. insert an image &
we select it, we switch tab from Home), but don't switch from the
Review, just show the Table tab.

With this, it's possible to visit each change in the document when some
of the changes are outside the table and some of them are inside a
table.

Signed-off-by: Miklos Vajna <vmiklos@collabora.com>
Change-Id: Ie9ed05f206a408c7b00168eacc72b486a01ccaaf