fix(clipboard/write): Defer fetching and checking
Previously we were awaiting some of our operations - e.g. the fetch or
the _check - before calling navigator.clipboard.write. This worked in
most places, but unfortunately iOS seems to fail when we do this.

While it wasn't possible to tell from debugging whether it was a small
delay that caused the clipboard write to fail or whether awaiting some
promises cause the failure in-and-of, deferring the promises into the
clipboard item fixes the issue on iOS.

Doing this requires a bit of promise juggling - in particular, we need
to reject if we have a check failure to avoid writing to the clipboard.
Doing this, however, will cause a generic failure that we previously
interpreted as a permission error. To bypass this, we can assume that a
permission failure with failing checks is not actually a permission
failure.

Altogether, the code ends up more ugly but works on iOS web. Oh well.

Signed-off-by: Skyler Grey <skyler.grey@collabora.com>
Change-Id: If20bb3cf715c40777395edd0cd431bcba5074217