run-wsdunit: add a mode that stops on first success
One way to investigate an unstable test is to have logs for both a
failing and passing case. If it takes several tries to find a passing
case, then this new mode can help, so if you run a test 10 times, then
it won't overwrite a passing run with a next, potentially again failing
run.

Signed-off-by: Miklos Vajna <vmiklos@collabora.com>
Change-Id: I22fe78016021ea4308db2b7d0dc4f484b06a6a82