feat(video): Playback on click
Previously videos would start playback as soon as they had loaded,
however this gives no control to the presenter which isn't ideal.

This commit makes it so videos no longer autoplay. Instead, they must be
manually started by clicking on them.

I'm not sure how to handle videos when the presenter is only using a
keyboard/clicker/etc. so I haven't. I figure that's a point for later
discussion.

Similarly, we may want to consider adding some buttons in the corner
of the presentation (perhaps when hovered) that allow you to control
the slides, since as a fullscreen video will prevent you from clicking
to advance. That said, this isn't really a video problem: shapes with
actions would do the same thing, and it is already a potential problem
that only advancing is possible via clicking, not going back; the
expectation as-of-now seems to be "use a keyboard/clicker/etc. to
advance slides"

This doesn't override swipe gestures to advance slides/etc. on
touchscreens, however you are able to tap to play the video.

Adapt patch to LibreOffice Online.

Signed-off-by: Skyler Grey <skyler.grey@collabora.com>
Change-Id: I483f42b6fee8f4d52909ece3b599670318ed650d