feat(video): Pause playback with a second click
Previously, clicking the video again would restart the video. We likely
more-commonly want to pause the video, for example if we would like to
halt the audio while we talk.

Notably this means that playback should resume from where we left off,
except where we have reached the end of the video in which case we still
want to restart from the top. To keep track of that, we'll need to also
listen to the "ended" event so that we can differentiate it from pausing
for another reason.

Signed-off-by: Skyler Grey <skyler.grey@collabora.com>
Change-Id: I5e7ec6ea3518987bf8e966a77bbd55f66a99d2e1