fix: Remove unsupported video warning on audio
Previously, we could upload any multimedia files (video or audio-only)
but audio files would show an "unsupported video" warning when played in
the editor in Chromium-based browsers.

That turns out to be because we test for "video.webkitDecodedFrameCount"
which is a Chromium-specific property to see whether any video frames
have been decoded. If the property exists and is 0 we assume that, since
as no video has been decoded, the file is corrupted or unsupported. As
audio doesn't have any video frames to decode, we will always see it as
undecodeable and therefore always corrupted.

To fix this we can check webkitAudioDecodedByteCount to see if we have
decoded any audio. Since as some videos can have no audio track we still
need to check video as well, however corrupted or unsupported videos
will not read any audio *or* video frames.

Signed-off-by: Skyler Grey <skyler.grey@collabora.com>
Change-Id: Id223a1893743ce1b3d41fd954f6a23ce3f8f51a3