wsd: abort conversion on save-as errors
This can happen when the format is not
supported or is invalid, or some other
conversion error has been encountered.

Without this, we wait a while until
the various time-outs kick in..

We also close the socket after
sending the error, since there
is nothing more to do after finishing
conversion.

Change-Id: I3d9e9cda539eec80ea567d7f6579be83e17d7679
Signed-off-by: Ashod Nakashian <ashod.nakashian@collabora.co.uk>