fix(slideshow): Prevent memory leak
Previously, we forgot to close our rendered bitmaps of slides after
rendering, instead leaving them for the garbage collector to deal with.

As per the MDN docs for transferToImageBitmap:

> If your goal is to pass the ImageBitmap to other web APIs which do
> not consume it - for example, CanvasRenderingContext2D.drawImage()
> - then you should close it when you're done with it by calling
> ImageBitmap.close(). Don't simply drop the JavaScript reference to the
> ImageBitmap; doing so will keep its graphics resource alive until the
> next time the garbage collector runs

(https://developer.mozilla.org/en-US/docs/Web/API/OffscreenCanvas/transferToImageBitmap)

This didn't tend to become a huge problem - obviously it's not ideal but
we weren't usually rendering slides that much and the garbage collector
doesn't run *that* infrequently - except in one case... When playing
video, we rerender the slide for each frame of the animation. As videos
take longer to finish than regular animations, we could get a very
noticable increase in memory. Further, as we were doing lots of work,
it seems that modern browsers on lower-powered machines would decide to
slow down garbage collection - meaning our old frames would not be
cleaned up until the video was no longer playing!

Signed-off-by: Skyler Grey <skyler.grey@collabora.com>
Change-Id: Ib4d2d43e7518064b71e60365d99ab52dd8170f78