wsd: admin: no need to keep expired documents around
We only care about the stats and the serialized
history of expired documents. Neither of these
change once a document has expired, so we
simply capture them and discard the Document
instance that has expired.

This should reduce memory pressure/fragmentation
at the expense of keeping the serialized history
string indefinitely.

Change-Id: I26b407be3f08ad808f88398cd256bd78f7e6fbe6
Signed-off-by: Ashod Nakashian <ashod.nakashian@collabora.co.uk>